Refactor to alleviate public/dist watching
