+++ lib/assets.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const { resolve } = require('path');
+const { notifyClient } = require('./watch');
 
 
 function asset(opts) {
@@ -13,8 +14,6 @@ function asset(opts) {
 
 
 function makeAssetsMap(outputs, opts) {
-  const map = {};
-
   Object.entries(outputs).forEach(([path, value]) => {
     const key = value.entryPoint
       ? resolve(value.entryPoint)
@@ -22,19 +21,30 @@ function makeAssetsMap(outputs, opts) {
 
     const hashedPath = path.replace(/^public/, '');
 
-    map[key] = {};
+    if (!opts.assets.map[key]) {
+      opts.assets.map[key] = {};
+    }
+
+    if (opts.watch) {
+      const fileSize = fs.statSync(resolve(path)).size;
+
+      if (opts.assets.map[key].size !== fileSize) {
+        console.log('changed now notify!!!!!', path, key);
+        notifyClient(hashedPath);
+      }
+
+      opts.assets.map[key].size = fileSize;
+    }
 
     if (opts.assets.inline.includes(key)) {
       const cssPath = resolve(__base, '../public', hashedPath.replace(/^\//, ''));
-      map[key].content = fs.readFileSync(cssPath, 'utf8');
+      opts.assets.map[key].content = fs.readFileSync(cssPath, 'utf8');
     }
 
-    map[key].path = key.match(/\.svg$/)
+    opts.assets.map[key].path = key.match(/\.svg$/)
       ? hashedPath
       : `${opts.assets.origin}${hashedPath}`;
   });
-
-  opts.assets.map = map;
 }
