+++ lib/index.js
@@ -2,25 +2,17 @@ const fs = require('fs');
 const callsites = require('callsites');
 const traversy = require('traversy');
 const beard = require('beard');
-const WebSocket = require('ws');
 const config = require('./config');
 const esbuild = require('./esbuild');
-const { dirname, basename, resolve, extname } = require('path');
+const { dirname, resolve, extname } = require('path');
 const { fileExtRegexStr, makeScopedClass, isDev } = require('./utils');
 const { blockTypes, extractBlocks, addHmrToBundles, writeEntryFile } = require('./bundling');
 const regex = new RegExp(fileExtRegexStr, 'g');
+const { startWatching, notifyClient } = require('./watch');
 
 
 exports.balm = (options = {}) => {
   const hashes = {};
-  const hasBundled = {};
-
-
-  let _socket;
-  let _timer;
-  let _changes = [];
-
-
   const opts = config(options);
   const {
     root, blocksDir, loadHandles,
@@ -44,7 +36,13 @@ exports.balm = (options = {}) => {
     }
 
     if (watch) {
-      startWatching();
+      startWatching(opts);
+
+      fs.watch(root, { recursive: true }, (_, path) => {
+        if (extname(path) === '.balm') {
+          bundleFile(resolve(root, path));
+        }
+      });
     }
 
     writeEntryFile('css', blocksDir);
@@ -79,50 +77,6 @@ exports.balm = (options = {}) => {
         }
       }
     });
-  }
-
-
-  function startWatching() {
-    console.log('Balm watching...');
-
-    _socket = new WebSocket.Server({ port });
-
-    fs.watch('./public/dist', (_, path) => {
-      console.log('THIS CAUSES A BUG ON FILES THAT HAVE ALREADY BEEN CREATED');
-
-      if (!hasBundled[path]) {
-        console.log('File changed for HMR', path);
-        notifyClient(path);
-        hasBundled[path] = true;
-      }
-    });
-
-    fs.watch(root, { recursive: true }, (_, path) => {
-      if (extname(path) === '.balm') {
-        bundleFile(resolve(root, path));
-      }
-    });
-  }
-
-
-  function notifyClient(path) {
-    clearTimeout(_timer);
-
-    _changes.push(basename(path));
-    _changes = [...new Set(_changes)];
-
-    _timer = setTimeout(() => {
-      if (!_changes.length) return;
-
-      _socket.clients.forEach((client) => {
-        if (client.readyState === WebSocket.OPEN) {
-          console.log('Balm notifying client of changes', _changes);
-          client.send(JSON.stringify(_changes));
-        }
-      });
-
-      _changes = [];
-    }, 50);
   }
